fix: insert lower and upper values as double

+++ lib/pages/settings_page/pages/alarms_page/alarm_config_page/parameter_widget.dart
@@ -267,8 +267,8 @@ class _ParameterWidgetState extends State<ParameterWidget> {
               onPressed: () {
                 context.read<DatabaseProvider>().updateAlarmParameter(
                     widget.parameterData.copyWith(
-                        lowerValue: Value(startValue.toInt()),
-                        upperValue: Value(endValue.toInt())));
+                        lowerValue: Value(startValue),
+                        upperValue: Value(endValue)));
 
                 setState(() {
                   _sliderValues = SfRangeValues(startValue, endValue);
